chatbot/requirements.txt
fastapi
python-dotenv
pymongo
python-socketio[asgi]
google-cloud-dialogflow
uvicorn
